Move all integrationtest-application components one folder level up
