--- _integration_tests/application/.gitignore
@@ -1,1 +1,0 @@
-!consumer_api_core
