+++ _integration_tests/config/test/consumer_api_core/consumer_api_iam_service.json
