+++ _integration_tests/config/test/iam/claim_service/claims/default_namespace/entity_types/user.json
