+++ _integration_tests/config/test/security/symmetric_encryption_service.json
