+++ _integration_tests/config/test/services/shell.json
