+++ _integration_tests/package.json
@@ -11,8 +11,8 @@
     "tests"
   ],
   "scripts": {
-    "start": "cross-env NODE_ENV=test CONFIG_PATH=application/config node application/index.js",
-    "test": "cross-env NODE_ENV=test CONFIG_PATH=application/config mocha test/**/*.js --exit"
+    "start": "cross-env NODE_ENV=test CONFIG_PATH=config node index.js",
+    "test": "cross-env NODE_ENV=test CONFIG_PATH=config mocha test/**/*.js --exit"
   },
   "dependencies": {
     "@essential-projects/bootstrapper": "^1.0.0",
