+++ _integration_tests/test/consumer_api_client/user_tasks/finish_user_task.js
@@ -2,7 +2,7 @@
 
 const should = require('should');
 
-const testSetup = require('../../../application/test_setup');
+const testSetup = require('../../../test_setup');
 
 const testTimeoutMilliseconds = 5000;
